% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{choose_db_path}
\alias{choose_db_path}
\title{Choose where to look for the generation and usage db.}
\usage{
choose_db_path(
  location = "default",
  path = system.file("extdata/generation_and_usage.sqlite3", package =
    "radiantNetDashboard"),
  filename = "generation_and_usage.sqlite3"
)
}
\arguments{
\item{location}{One of
\itemize{
\item "default": The internal example DB.
\item "path": Whatever is specified in the \code{path} param.
\item "user": The users platform specific data dir, see
\code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}}.
\item "site": The machine-wide platform specific data dir, see
\code{\link[rappdirs:site_data_dir]{rappdirs::site_data_dir()}}.
}}

\item{path}{The path to be used when \code{location} == "path".}

\item{filename}{The filename of the DB to be used for
\code{location \%in\% c("site", "user")}.}
}
\value{
A path in string form to the chosen DB.
}
\description{
Choose where to look for the generation and usage db.
}
