% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyProduction.R
\name{dailyProductionServer}
\alias{dailyProductionServer}
\title{Server for the Shiny module displaying the data for daily production.}
\usage{
dailyProductionServer(id, db_pool)
}
\arguments{
\item{id}{See same argument in \code{\link[shiny:moduleServer]{shiny::moduleServer()}}.}

\item{db_pool}{Pool of connections to the generation & usage database.}
}
\description{
Server for the Shiny module displaying the data for daily production.
}
