% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_agg_df_date}
\alias{get_agg_df_date}
\title{Get all the daily aggregated data on the same level of time as a given date.
E.g.: All days in the same month as 2024-12-01, or all days in the same year
as 2024-01-01.}
\usage{
get_agg_df_date(db_conn, date, format = "\%Y-\%m-\%d", level = "month")
}
\arguments{
\item{db_conn}{Connection to the generation & usage database.}

\item{date}{String describing the date for which data should be retrieved.
parsed into a date object internally.}

\item{format}{Format of the \code{date} string, for parsing it to a date object.
Parsing is done via \code{\link[base:strptime]{base::strptime()}}.}

\item{level}{The level of time for which unique dates should be retrieved.
One of "total", "year", or "month".}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of the \code{daily_aggregated} table with only the
data for the specified date.
}
\description{
Get all the daily aggregated data on the same level of time as a given date.
E.g.: All days in the same month as 2024-12-01, or all days in the same year
as 2024-01-01.
}
