% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{get_app_with_pool}
\alias{get_app_with_pool}
\title{Generate a list containing server and UI components for the dashboard bundled
with a connection pool to the app's database.}
\usage{
get_app_with_pool()
}
\description{
Generate a list containing server and UI components for the dashboard bundled
with a connection pool to the app's database.
}
