% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config_db_path}
\alias{get_config_db_path}
\title{Get the configured path to the generation and usage DB, as determined by
\code{choose_db_path}.}
\usage{
get_config_db_path()
}
\description{
Get the configured path to the generation and usage DB, as determined by
\code{choose_db_path}.
}
