% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_pkg_info}
\alias{get_pkg_info}
\title{Extract package information from the citation.}
\usage{
get_pkg_info(pkg_name = "radiantNetDashboard")
}
\arguments{
\item{pkg_name}{A string giving the name of the package for which the info
should be retrieved.}
}
\description{
Extract package information from the citation.
}
