% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_raw_data_df}
\alias{get_raw_data_df}
\title{Load the \code{raw_data} table into memory as a tibble.}
\usage{
get_raw_data_df(db_conn)
}
\arguments{
\item{db_conn}{Connection to the generation & usage database.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of the \code{raw_data} table.
}
\description{
Load the \code{raw_data} table into memory as a tibble.
}
