% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_raw_data_df_date}
\alias{get_raw_data_df_date}
\title{Load the raw data df filtered for a given date into memory as a tibble.}
\usage{
get_raw_data_df_date(db_conn, date, format = "\%Y-\%m-\%d")
}
\description{
Load the raw data df filtered for a given date into memory as a tibble.
}
