% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{radiant_net_dashboard}
\alias{radiant_net_dashboard}
\title{Run the dashboard.}
\usage{
radiant_net_dashboard(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\description{
Run the dashboard.
}
