% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{radiantNetDashboard}
\alias{radiantNetDashboard}
\title{Run the dashboard.}
\usage{
radiantNetDashboard(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\description{
Run the dashboard.
}
