% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{report_panel}
\alias{report_panel}
\title{Display a navigation panel containing reports of daily, monthly, yearly,
and total consumption/production data.}
\usage{
report_panel(title, db_pool, id_list)
}
\arguments{
\item{title}{Title of the entire navigation panel.}

\item{db_pool}{Pool of connections to the generation & usage database.}

\item{id_list}{Named list of ids connecting the UIs of each timeframe (daily,
monthly, etc.) to their respective servers.}
}
\description{
Display a navigation panel containing reports of daily, monthly, yearly,
and total consumption/production data.
}
