% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyProduction.R
\name{safeDateInput}
\alias{safeDateInput}
\title{Create a dateInput with only those dates enabled for which there is data in
a DB.}
\usage{
safeDateInput(inputId, db_pool)
}
\arguments{
\item{inputId}{See same argument in \code{\link[shiny:dateInput]{shiny::dateInput()}}.}

\item{db_pool}{Pool of connections to the generation & usage database.}
}
\description{
Create a dateInput with only those dates enabled for which there is data in
a DB.
}
