% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedProduction.R
\name{summarize_agg_data}
\alias{summarize_agg_data}
\title{Summarize aggregated daily data over a level of time.}
\usage{
summarize_agg_data(data, level)
}
\arguments{
\item{data}{Dataframe to summarize. Expected to contain a column of the time
level \emph{across} (i.e. one level below what is passed to \code{level}) which should
be summarized and some columns starting with "kwh_" which will be
summarized.}

\item{level}{The time level \emph{on} which we want to summarize. Either "month",
"year", or "total".}
}
\description{
Summarize aggregated daily data over a level of time.
}
