% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{wip_box}
\alias{wip_box}
\title{Display a box indicating this content is currently under construction.}
\usage{
wip_box()
}
\description{
Display a box indicating this content is currently under construction.
}
