% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{ensure_safe_date}
\alias{ensure_safe_date}
\title{Ensure a date is valid, else return the latest date with data.}
\usage{
ensure_safe_date(date, db_conn)
}
\description{
Ensure a date is valid, else return the latest date with data.
}
