% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_raw_data_stat_date}
\alias{get_raw_data_stat_date}
\title{Get date matching a statistic with data from the raw_data table.}
\usage{
get_raw_data_stat_date(db_conn, stat_fun = min, fun_args = list(na.rm = TRUE))
}
\description{
Get date matching a statistic with data from the raw_data table.
}
